+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/asset_publisher.ftl
@@ -1,4 +1,4 @@
-<#assign pageCounts = dataFactory.getSequence(dataFactory.maxAssetPublisherPageCount) />
+<#assign pageCounts = counterDataFactory.getSequence(dataFactory.maxAssetPublisherPageCount) />
 
 <#list pageCounts as pageCount>
 	<#assign
