+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/blogs.ftl
@@ -20,7 +20,7 @@ ${dataFactory.toInsertSQL(blogDataFactory.newUserNotificationDeliveryModel("com_
 		_entry=blogsEntryModel
 	/>
 
-	<#assign mbRootMessageId = dataFactory.getCounterNext() />
+	<#assign mbRootMessageId = counterDataFactory.getCounterNext() />
 
 	<@insertMBDiscussion
 		_classNameId=blogDataFactory.blogsEntryClassNameId
@@ -28,7 +28,7 @@ ${dataFactory.toInsertSQL(blogDataFactory.newUserNotificationDeliveryModel("com_
 		_groupId=groupId
 		_maxCommentCount=blogDataFactory.maxBlogsEntryCommentCount
 		_mbRootMessageId=mbRootMessageId
-		_mbThreadId=dataFactory.getCounterNext()
+		_mbThreadId=counterDataFactory.getCounterNext()
 	/>
 
 	${dataFactory.toInsertSQL(dataFactory.newSubscriptionModel(blogsEntryModel))}
